requirements.txt
fastapi
uvicorn[standard]
pydantic
pandas
numpy
scikit-learn
xgboost
mlflow
boto3
evidently
prometheus-client
python-dotenv
requests
